# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PixelAid API",
    "summary": "HTTP API for the technofest2026 PixelAid workspace.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/public/cases/{case_id}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Public Case",
        "operationId": "public_case_api_public_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseBrief"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/specialists/{specialist_id}/cases": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Public Specialist Cases",
        "operationId": "public_specialist_cases_api_public_specialists__specialist_id__cases_get",
        "parameters": [
          {
            "name": "specialist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Specialist Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseBrief"
                  },
                  "title": "Response Public Specialist Cases Api Public Specialists  Specialist Id  Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Me",
        "operationId": "me_api_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/profile": {
      "patch": {
        "tags": [
          "me"
        ],
        "summary": "Patch Profile",
        "operationId": "patch_profile_api_me_profile_patch",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/onboarding-complete": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Onboarding Complete",
        "operationId": "onboarding_complete_api_me_onboarding_complete_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/specialists": {
      "get": {
        "tags": [
          "specialists"
        ],
        "summary": "Specialists",
        "operationId": "specialists_api_specialists_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicSpecialist"
                  },
                  "title": "Response Specialists Api Specialists Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/specialists/{specialist_id}/cases": {
      "get": {
        "tags": [
          "specialists"
        ],
        "summary": "Specialist Cases",
        "operationId": "specialist_cases_api_specialists__specialist_id__cases_get",
        "parameters": [
          {
            "name": "specialist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Specialist Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseCard"
                  },
                  "title": "Response Specialist Cases Api Specialists  Specialist Id  Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Case Brief",
        "operationId": "case_brief_api_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseBrief"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CaseBrief": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "specialist_id": {
            "type": "string",
            "title": "Specialist Id"
          },
          "specialist_name": {
            "type": "string",
            "title": "Specialist Name"
          },
          "patient_name": {
            "type": "string",
            "title": "Patient Name"
          },
          "patient_age": {
            "type": "integer",
            "title": "Patient Age"
          },
          "patient_gender": {
            "type": "string",
            "title": "Patient Gender"
          },
          "chief_complaint": {
            "type": "string",
            "title": "Chief Complaint"
          },
          "triage_note": {
            "type": "string",
            "title": "Triage Note"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "condition_badge": {
            "type": "string",
            "title": "Condition Badge"
          },
          "estimated_duration_minutes": {
            "type": "integer",
            "title": "Estimated Duration Minutes"
          },
          "learning_objectives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Learning Objectives"
          }
        },
        "type": "object",
        "required": [
          "id",
          "specialist_id",
          "specialist_name",
          "patient_name",
          "patient_age",
          "patient_gender",
          "chief_complaint",
          "triage_note",
          "difficulty",
          "condition_badge",
          "estimated_duration_minutes"
        ],
        "title": "CaseBrief"
      },
      "CaseCard": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "specialist_id": {
            "type": "string",
            "title": "Specialist Id"
          },
          "specialist_name": {
            "type": "string",
            "title": "Specialist Name"
          },
          "patient_name": {
            "type": "string",
            "title": "Patient Name"
          },
          "patient_age": {
            "type": "integer",
            "title": "Patient Age"
          },
          "patient_gender": {
            "type": "string",
            "title": "Patient Gender"
          },
          "chief_complaint": {
            "type": "string",
            "title": "Chief Complaint"
          },
          "triage_note": {
            "type": "string",
            "title": "Triage Note"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "condition_badge": {
            "type": "string",
            "title": "Condition Badge"
          },
          "estimated_duration_minutes": {
            "type": "integer",
            "title": "Estimated Duration Minutes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "specialist_id",
          "specialist_name",
          "patient_name",
          "patient_age",
          "patient_gender",
          "chief_complaint",
          "triage_note",
          "difficulty",
          "condition_badge",
          "estimated_duration_minutes"
        ],
        "title": "CaseCard"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeResponse": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "progress": {
            "$ref": "#/components/schemas/ProgressResponse"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "progress"
        ],
        "title": "MeResponse"
      },
...
      "Profile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "is_anonymous": {
            "type": "boolean",
            "title": "Is Anonymous",
            "default": false
          },
          "onboarding_completed": {
            "type": "boolean",
            "title": "Onboarding Completed",
            "default": false
          },
          "xp": {
            "type": "integer",
            "title": "Xp",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "Profile"
      },
      "ProfileUpdate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "title": "ProfileUpdate"
      },
      "ProgressResponse": {
        "properties": {
          "total_xp": {
            "type": "integer",
            "title": "Total Xp",
            "default": 0
          },
          "level": {
            "type": "integer",
            "title": "Level",
            "default": 1
          },
          "completed_cases": {
            "type": "integer",
            "title": "Completed Cases",
            "default": 0
          },
          "average_best_score": {
            "type": "number",
            "title": "Average Best Score",
            "default": 0
          }
        },
        "type": "object",
        "title": "ProgressResponse"
      },
      "PublicSpecialist": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "case_count": {
            "type": "integer",
            "title": "Case Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "icon",
          "status"
        ],
        "title": "PublicSpecialist"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}